Add function SecureJoinVFSoptionalScope() for optionally disabling chroot behaviour
When chroot behaviour is disabled, the provided root path is not treated
as the root of the filesystem. Instead, an error is returned when a path
component ".." is resolved from the root directory. An error is also
returned when a traversed symlink has a target that is an absolute path.

Signed-off-by: Erik Sjölund <erik.sjolund@gmail.com>